:building_construction: Move Auth API Types
Move Auth API Types #133
