+++ api/auth/types.ts
@@ -37,3 +37,30 @@ export type AuthError = AxiosError<{
   message?: AuthErrorData;
   data?: AuthErrorData;
 }>;
+
+export interface SignUpParams {
+  username: string;
+  password: string;
+  nickName: string;
+  phoneNumber: string;
+  gender: 'male' | 'female' | '';
+  birthDay: string;
+  termAgreement: boolean;
+  privacyAgreement: boolean;
+  category: string[];
+}
+
+export interface LoginParams {
+  username: string;
+  password: string;
+}
+
+export interface UpdatePasswordParams {
+  username: string;
+  toBePassword: string;
+}
+
+export interface ChekMemberParams {
+  username: string;
+  phoneNumber: string;
+}
